feat: add tryExceptional method to Result class
The tryExceptional method is added to the Result class. This method allows performing an action in a try-catch block and returning a Result object. If an exception is caught, a failure result with an error containing the throwable is constructed. This method is useful for handling exceptions in a functional way.